Updated full dataset model, updated lasso function legend title handling
The full dataset model has a simpler code and is more comprehensive - included ALL variables with the exception of a handful that were obviously misplaced variables (like pop 18 over included in the disability group, which I checked, was more in line with the actual population over 18 variable).

I also updated the legend title of the lasso function to indicate whether these are group labels or feature labels.