Static page is fully dynamic, indicators inline with dynamic page
I made several updates, indicators on both pages are not the same, the main page they are scaled by population now, the bar width and color is determied by the relative position of the weighted average in comparison to the min and max CAs. The percentage represents the actual rate (I changed that too for the dynamic page, not it displays percentage).

the year now affects safest and riskiest CAs, it changes the crime rates as well.

Comparison bars are still static, these will need to be modified to dynamically display the CAs that are determined by the comparison Chart